tests/wpt/tests/css/css-text/parsing/text-wrap-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Module Test: parsing text-wrap with valid values</title>
<link rel="author" title="Tim Nguyen" href="https://github.com/nt1m">
<link rel="help" href="https://drafts.csswg.org/css-text-4/#text-wrap">
<meta name="assert" content="text-wrap supports the full grammar 'wrap | nowrap | balance | stable | pretty'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("text-wrap", "wrap");
test_valid_value("text-wrap", "nowrap");
test_valid_value("text-wrap", "balance");
test_valid_value("text-wrap", "stable");
test_valid_value("text-wrap", "pretty");
test_valid_value("text-wrap", "initial");
test_valid_value("text-wrap", "inherit");
test_valid_value("text-wrap", "unset");
test_valid_value("text-wrap", "revert");
test_valid_value("text-wrap", "revert-layer");
</script>
</body>
</html>
